requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.5.2; python_version >= '3.7'
certifi==2022.6.15; python_full_version >= '3.6.0'
charset-normalizer==2.1.0; python_full_version >= '3.6.0'
coreapi==2.3.3
coreschema==0.0.4
django-environ==0.9.0
django-filter==22.1
django==4.0.6
djangorestframework==3.13.1
drf-yasg==1.21.3
gunicorn==20.1.0
idna==3.3; python_version >= '3.5'
inflection==0.5.1; python_version >= '3.5'
itypes==1.2.0
jinja2==3.1.2; python_version >= '3.7'
markupsafe==2.1.1; python_version >= '3.7'
packaging==21.3; python_full_version >= '3.6.0'
psycopg2==2.9.3
pyparsing==3.0.9; python_full_version >= '3.6.8'
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1; python_version >= '3.7' and python_version < '4'
ruamel.yaml.clib==0.2.6; python_version < '3.11' and platform_python_implementation == 'CPython'
ruamel.yaml==0.17.21; python_version >= '3'
setuptools==63.2.0; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.2; python_version >= '3.5'
uritemplate==4.1.1; python_full_version >= '3.6.0'
urllib3==1.26.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5' and python_version < '4'
